+++ stationary_models_with_an_eye _on_quasi-stationary/Bernoulli_equation.cpp
@@ -1,0 +1,1 @@
+#include "Bernoulli_equation.h"
