+++ stationary_models_with_an_eye _on_quasi-stationary/stationary_models_with_an_eye _on_quasi-stationary.cpp
@@ -18,7 +18,7 @@
 
 // Подключение класса для определения гидравлического сопротивления 
 #include "Hydraulic_resistance_coefficient.h"
-#include "Hydraulic_resistance_coefficient.cpp"
+
 
 
 // Используем простравнство имен std и pde_solvers
@@ -41,10 +41,8 @@ struct Pipiline_parameters {
 	double l; 
 };
 
-
-
 /// @brief Oil_parameters - Структура парметров нефти
-/// @param ro - плотность нефти [кг/м3]
+/// @param ro - плотность нефти [кг/м^3]
 /// @param nu - кинематическая вязкость [Ст]
 /// @param p0 - давление в начале участка нефтепровода [Па]
 /// @param pl - давление в конце участка нефтепровода [Па]
@@ -58,79 +56,7 @@ struct Oil_parameters {
 };
 
 
-/// @brief Bernoulli_equation - класс для решения задач из блока 2 - Реализация стационарных моделей 
-/// с прицелом на квазистационар (Уравнение Бернулли)
-class Bernoulli_equation {
-public:
-	/// @brief pressure_p0 - Метод, рассчитывающий давление в начале участка нефтепровода [МПа]
-	/// @param pipiline_parameters_XX - структура парметров трубопровода
-	/// @param oil_parameters_XX - структура парметров нефти
-	/// @param lambda - коэффициент гидравлического сопротивления
-	/// @param v - скорость течения нефти в системе СИ
-	/// @param d - внутренний диаметр трубы в системе СИ
-	/// @return oil_parameters_XX.p0 - давление в начале участка нефтепровода [МПа]
-	double pressure_p0(Pipiline_parameters pipiline_parameters_XX, Oil_parameters oil_parameters_XX, double lambda, double v, double d) {
-		oil_parameters_XX.p0 = (oil_parameters_XX.ro * 9.81) * (oil_parameters_XX.pl * 1000000 / (oil_parameters_XX.ro * 9.81) - pipiline_parameters_XX.z0 + pipiline_parameters_XX.zl + (((lambda * pipiline_parameters_XX.l * 1000) / d) * pow(v, 2)) / (2 * 9.81)) * 0.000001;
-		return oil_parameters_XX.p0;
-	}
-	/// @brief diameter - метод, рассчитывающий внутренний диаметр трубы
-	/// @param pipiline_parameters_XX - структура парметров трубопровода
-	/// @return d - внутренний диаметр трубы [м]
-	double diameter(Pipiline_parameters pipiline_parameters_XX) {
-		double d = (pipiline_parameters_XX.D - 2 * pipiline_parameters_XX.ds) / 1000;
-		return d;
-	}
-	
-	/// @brief  relative_roughness - метод, рассчитывающий относительную эквивалентная шероховатость
-	/// @param pipiline_parameters_XX - структура парметров трубопровода
-	/// @param d - внутренний диаметр трубы [м]
-	/// @return e oтносительная эквивалентная шероховатость
-	double relative_roughness(Pipiline_parameters pipiline_parameters_XX, double d) {
-		double e = (pipiline_parameters_XX.delta / 1000) / d;
-		return e;
-	}
 
-	/// @brief reynolds_number - метод, рассчитывающий число Рейнольдса, где nu переведено в систему СИ
-	/// @param oil_parameters_XX - структура парметров нефти
-	/// @param v - cкорость течения нефти в системе СИ
-	/// @param d - внутренний диаметр трубы в системе СИ
-	/// @return Re - число Рейнольдса
-	double reynolds_number(Oil_parameters oil_parameters_XX, double v, double d) {
-		double Re = v * d / (oil_parameters_XX.nu * 0.000001);
-		return Re;
-	}
-	
-	/// @brief speed_flow - метод, рассчитывающий скорость по заданному расходу нефти
-	/// @param oil_parameters_XX - структура парметров нефти
-	/// @param d - внутренний диаметр трубы [м]
-	/// @return v - cкорость течения нефти в системе СИ
-	double speed_flow(Oil_parameters oil_parameters_QP, double d) {
-		double v = (4 * oil_parameters_QP.Q / 3600) / (3.14 * pow(d, 2));
-		return v;
-	}	
-
-	/// @brief speed_pressure - метод, рассчитывающий скорость по давлению в задаче PP
-	/// @param pipiline_parameters_XX - структура парметров трубопровода
-	/// @param oil_parameters_XX - структура парметров нефти
-	/// @param lambda - коэффициент гидравлического сопротивления
-	/// @param d - внутренний диаметр трубы в системе СИ
-	/// @param oil_parameters_XX.p0 - давление в начале участка нефтепровода [МПа]
-	/// @param oil_parameters_XX.pl - давление в конце участка нефтепровода [МПа]
-	/// @return v - cкорость течения нефти в системе СИ	
-	double speed_pressure(Pipiline_parameters pipiline_parameters_XX, Oil_parameters oil_parameters_XX, double lambda, double d) {
-		double v = pow((2 * 9.81 * d / pipiline_parameters_XX.l / 1000 * ((oil_parameters_XX.p0  - oil_parameters_XX.pl) * 1000000 / (oil_parameters_XX.ro * 9.81) + pipiline_parameters_XX.z0 - pipiline_parameters_XX.zl) / lambda), 0.5);
-		return v;
-	}
-
-	/// @brief volume_flow - метод, рассчитывающий объемный расход
-	/// @param v - скорость течения нефти в системе СИ
-	/// @param d - внутренний диаметр трубы [м]
-	/// @return Q - объемный расход [м^3/ч]
-	double volume_flow(double v, double d) {
-		double Q = 3.14 * pow(d, 2) * v * 3600 / 4;
-		return Q;
-	}
-};
 
 /// @brief Iterative_solutions - класс для решения задач из блока 2 численными методами
 class Iterative_solutions {
